requirements.txt
torch==1.11.0
torchvision==0.12.0
albumentations==1.2.0
numpy
pandas
matplotlib
tqdm
pyyaml